+++ src/config/keys/parse.rs
@@ -29,6 +29,7 @@ pub(in crate::config) struct KeysConfigOverride {
     pub(super) copy_path: Option<KeyConfigOverride>,
     pub(super) search_folders: Option<KeyConfigOverride>,
     pub(super) search_files: Option<KeyConfigOverride>,
+    pub(super) filter_directory: Option<KeyConfigOverride>,
     pub(super) zoxide: Option<KeyConfigOverride>,
     pub(super) shell: Option<KeyConfigOverride>,
     pub(super) open: Option<KeyConfigOverride>,